Repair text
Repaire some text due to conruption

+++ src/main/java/fr/erinagroups/erinium/procedures/BushBlockIsPlacedByProcedure.java
@@ -37,17 +37,16 @@ public class BushBlockIsPlacedByProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if (!((world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == Blocks.STONE
-				|| (world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == Blocks.DIRT
-				|| (world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == Blocks.GRASS_BLOCK
-				|| (world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == Blocks.GRASS_PATH
-				|| (world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == Blocks.MYCELIUM
-				|| (world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == Blocks.COARSE_DIRT
-				|| (world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == Blocks.PODZOL)) {
+		if (!((world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == Blocks.STONE
+				|| (world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == Blocks.DIRT
+				|| (world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == Blocks.GRASS_BLOCK
+				|| (world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == Blocks.GRASS_PATH
+				|| (world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == Blocks.MYCELIUM
+				|| (world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == Blocks.COARSE_DIRT
+				|| (world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == Blocks.PODZOL)) {
 			if (world instanceof World) {
-				Block.spawnDrops(world.getBlockState(new BlockPos((int) x, (int) y, (int) z)), (World) world,
-						new BlockPos((int) x, (int) y, (int) z));
-				world.destroyBlock(new BlockPos((int) x, (int) y, (int) z), false);
+				Block.spawnDrops(world.getBlockState(new BlockPos(x, y, z)), (World) world, new BlockPos(x, y, z));
+				world.destroyBlock(new BlockPos(x, y, z), false);
 			}
 		}
 	}
