+++ src/main/java/fr/erinagroups/erinium/procedures/SciningCactusAdditionalGenerationConditionProcedure.java
@@ -36,8 +36,8 @@ public class SciningCactusAdditionalGenerationConditionProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if ((world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == SciningSandBlock.block) {
-			world.setBlockState(new BlockPos((int) x, (int) y, (int) z), SciningCactusBlock.block.getDefaultState(), 3);
+		if ((world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == SciningSandBlock.block) {
+			world.setBlockState(new BlockPos(x, y, z), SciningCactusBlock.block.getDefaultState(), 3);
 			return true;
 		}
 		return false;
