+++ src/main/java/fr/erinagroups/erinium/procedures/SciningCactusNeighbourBlockChangesProcedure.java
@@ -37,15 +37,14 @@ public class SciningCactusNeighbourBlockChangesProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if ((world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == Blocks.AIR
-				|| !((world.getBlockState(new BlockPos((int) (x - 1), (int) y, (int) z))).getBlock() == Blocks.AIR)
-				|| !((world.getBlockState(new BlockPos((int) (x + 1), (int) y, (int) z))).getBlock() == Blocks.AIR)
-				|| !((world.getBlockState(new BlockPos((int) x, (int) y, (int) (z - 1)))).getBlock() == Blocks.AIR)
-				|| !((world.getBlockState(new BlockPos((int) x, (int) y, (int) (z + 1)))).getBlock() == Blocks.AIR)) {
+		if ((world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == Blocks.AIR
+				|| !((world.getBlockState(new BlockPos(x - 1, y, z))).getBlock() == Blocks.AIR)
+				|| !((world.getBlockState(new BlockPos(x + 1, y, z))).getBlock() == Blocks.AIR)
+				|| !((world.getBlockState(new BlockPos(x, y, z - 1))).getBlock() == Blocks.AIR)
+				|| !((world.getBlockState(new BlockPos(x, y, z + 1))).getBlock() == Blocks.AIR)) {
 			if (world instanceof World) {
-				Block.spawnDrops(world.getBlockState(new BlockPos((int) x, (int) y, (int) z)), (World) world,
-						new BlockPos((int) x, (int) y, (int) z));
-				world.destroyBlock(new BlockPos((int) x, (int) y, (int) z), false);
+				Block.spawnDrops(world.getBlockState(new BlockPos(x, y, z)), (World) world, new BlockPos(x, y, z));
+				world.destroyBlock(new BlockPos(x, y, z), false);
 			}
 		}
 	}
