+++ src/main/java/fr/erinagroups/erinium/procedures/SciningCactusUpdateTickProcedure.java
@@ -36,14 +36,14 @@ public class SciningCactusUpdateTickProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if (!((world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == SciningCactusBlock.block)) {
-			if ((world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z))).getBlock() == Blocks.AIR) {
-				world.setBlockState(new BlockPos((int) x, (int) (y + 1), (int) z), SciningCactusBlock.block.getDefaultState(), 3);
+		if (!((world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == SciningCactusBlock.block)) {
+			if ((world.getBlockState(new BlockPos(x, y + 1, z))).getBlock() == Blocks.AIR) {
+				world.setBlockState(new BlockPos(x, y + 1, z), SciningCactusBlock.block.getDefaultState(), 3);
 			}
 		} else {
-			if (!((world.getBlockState(new BlockPos((int) x, (int) (y - 2), (int) z))).getBlock() == SciningCactusBlock.block)) {
-				if ((world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z))).getBlock() == Blocks.AIR) {
-					world.setBlockState(new BlockPos((int) x, (int) (y + 1), (int) z), SciningCactusBlock.block.getDefaultState(), 3);
+			if (!((world.getBlockState(new BlockPos(x, y - 2, z))).getBlock() == SciningCactusBlock.block)) {
+				if ((world.getBlockState(new BlockPos(x, y + 1, z))).getBlock() == Blocks.AIR) {
+					world.setBlockState(new BlockPos(x, y + 1, z), SciningCactusBlock.block.getDefaultState(), 3);
 				}
 			}
 		}
